+++ pkg/analysis_server/lib/src/services/correction/executable_parameters.dart
@@ -73,9 +73,9 @@ class ExecutableParameters {
 
   static ExecutableParameters? forInvocation(
     AnalysisSessionHelper sessionHelper,
-    AstNode? invocation,
-  ) {
-    Element? element;
+    AstNode? invocation, {
+    Element? element,
+  }) {
     // This doesn't handle FunctionExpressionInvocation.
     if (invocation is Annotation) {
       element = invocation.element;